+++ client/package.json
@@ -11,6 +11,7 @@
     "@types/node": "^12.0.0",
     "@types/react": "^16.9.53",
     "@types/react-dom": "^16.9.8",
+    "@types/react-router-dom": "^5.1.7",
     "@types/react-slick": "^0.23.4",
     "add": "^2.0.6",
     "firebase": "^8.2.1",
@@ -19,6 +20,7 @@
     "react-dom": "^17.0.1",
     "react-easy-crop": "^3.3.1",
     "react-images-upload": "^1.2.8",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.1",
     "typescript": "^4.0.3",
     "web-vitals": "^0.2.4",
